Refactor Kind to Schedule.Kind and update usage
Moved the Kind type into Schedule as Schedule.Kind, updating all references and extensions accordingly. Deleted the standalone Kind.swift file and migrated related logic and error handling to Schedule.Kind. Schedule now conforms to Codable, and Schedule.Kind includes Codable and Sendable conformance.
